+++ src/main/java/br/com/schoolify/shoolify/repository/DiscProfTurmaRepository.java
@@ -3,5 +3,5 @@ package br.com.schoolify.shoolify.repository;
 import br.com.schoolify.shoolify.model.DiscProfTurma;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DiscProfTurmaRepository extends JpaRepository<DiscProfTurma, Integer> {
+public interface DiscProfTurmaRepository extends JpaRepository<DiscProfTurma, Long> {
 }
