+++ src/main/java/br/com/schoolify/shoolify/repository/TurmaPaiRepository.java
@@ -3,5 +3,5 @@ package br.com.schoolify.shoolify.repository;
 import br.com.schoolify.shoolify.model.TurmaPai;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TurmaPaiRepository extends JpaRepository<TurmaPai, Integer> {
+public interface TurmaPaiRepository extends JpaRepository<TurmaPai, Long> {
 }
